Updated pin search right after press and hold with fly to transition
