+++ TripBorderFrontend/src/components/trips/TripsItems.jsx
@@ -10,7 +10,6 @@ import ButtonPOISelection from './ButtonPOISelection';
 import ToggleDice from './ToggleDice';
 import TripSearchTools from './TripSearchTools';
 import ButtonGPSSearch from './ButtonGPSSearch';
-import ButtonPinSearch from './ButtonPinSearch';
 import CustomError from '../CustomError';
 import TripPlanningTools from './TripPlanningTools';
 
@@ -24,10 +23,10 @@ function TripsItems() {
 
   const getAPIStatus = () => {
     if (isLoading) {
-      return <span>Loading...</span>;
+      return <span>Loading</span>;
     }
     if (isFetching) {
-      return <span>Fetching...</span>;
+      return <span>Fetching</span>;
     }
 
     if (error) {
@@ -42,7 +41,6 @@ function TripsItems() {
       <div className='flex-col overflow-x-auto pt-1.5'>
         <div className='max-h-13'>
           <ButtonGPSSearch getNearbyPOIQueryTrigger={getNearbyPOIQueryTrigger} />
-          <ButtonPinSearch getNearbyPOIQueryTrigger={getNearbyPOIQueryTrigger} />
           <ButtonPOISelection reset={reset} isFetching={isFetching} />
           <ToggleDice poi={poi} />
           <CustomToggle
@@ -55,14 +53,15 @@ function TripsItems() {
             title='🏖️'
             component={<TripPlanningTools />}
           />
+          {getAPIStatus()}
         </div>
-        {getAPIStatus()}
       </div>
       <div>
         <CustomMap
           data={(poi) || null}
-          getPOIPhotosQueryResult={(getPOIPhotosQueryResult) || null}
+          getNearbyPOIQueryTrigger={getNearbyPOIQueryTrigger}
           getPOIPhotosQueryTrigger={getPOIPhotosQueryTrigger}
+          getPOIPhotosQueryResult={(getPOIPhotosQueryResult) || null}
         />
       </div>
     </div>
